third_party/blink/web_tests/platform/linux-chrome/external/wpt/fetch/stale-while-revalidate/fetch-sw.https-expected.txt

This is a testharness.js-based test.
[FAIL] Second fetch returns same response
  assert_equals: expected "vfuoexberllddzyszzdg" but got "axeuxcjvyvryyxnddemj"
Harness: the test ran to completion.
